Disable Parent Links
The JS script provided above allows to either disable the parent links or keep them clickable.

To disable the parent links call the dvcs_collapse_menu_module_submenus_on_mobile() function without passing any arguments (or pass false) to it. In this case clicking the parent item will only toggle the submenu.

Keep Parent Links Clickable
However, if you need to be able to toggle the submenus but at the same time the parent links still to be clickable then pass true as the argument.